+++ clid/main.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-__version__ = '0.7.0'
+__version__ = '0.6.3.1'
 
 import os
 import curses